Fix thread publication
